Solving several problems compiling one-core-api wrappers and dependencies. Work in prorgress

+++ wrappers/dependencies/userbase/misc/desktop.c
@@ -561,7 +561,7 @@ GetThreadDesktop(
     }
 
     return NtUserGetThreadDesktop(dwThreadId,
-                                  (DWORD)GetThreadConsoleDesktopRequest->ConsoleDesktop);
+                                 GetThreadConsoleDesktopRequest->ConsoleDesktop);
 }
